Modified rng and seed to be deterministic
Replaced the integer seed with a hash128 which gets created from user input. The hash is converted to the random state which is assigned to the static game random class. This class wraps Unity's random but ensures that it is deterministic (serperates unity's calls to random from script calls). There are now 2^128 possible seeds instead of 2^32.